+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Scripts.hs
@@ -1,8 +1,19 @@
 module Cardano.Ledger.Api.Scripts (
   module Cardano.Ledger.Api.Scripts.Data,
   EraScript (..),
+  ScriptHash,
+  CostModels (..),
+  ValidityInterval (..),
+
+  -- * Tools
+  module Cardano.Ledger.Api.Scripts.Tools,
 )
 where
 
+import Cardano.Ledger.Allegra.Scripts (ValidityInterval (..))
+import Cardano.Ledger.Alonzo.Scripts (CostModels (..))
+import Cardano.Ledger.Api.Era ()
 import Cardano.Ledger.Api.Scripts.Data
+import Cardano.Ledger.Api.Scripts.Tools
 import Cardano.Ledger.Core (EraScript (..))
+import Cardano.Ledger.Hashes (ScriptHash)