+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/Address.hs
@@ -1,0 +1,20 @@
+module Cardano.Ledger.Api.Tx.Address (
+  -- * Address
+  Addr (..),
+  BootstrapAddress (..),
+  serialiseAddr,
+  decodeAddr,
+  decodeAddrEither,
+  decodeAddrShort,
+  decodeAddrShortEither,
+  getNetwork,
+
+  -- * Reward Account
+  RewardAcnt (..),
+  mkRwdAcnt,
+  serialiseRewardAcnt,
+  deserialiseRewardAcnt,
+)
+where
+
+import Cardano.Ledger.Address