+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/AuxData.hs
@@ -1,6 +1,26 @@
 module Cardano.Ledger.Api.Tx.AuxData (
   EraTxAuxData (..),
+
+  -- * Shelley
+  ShelleyTxAuxData (..),
+  Metadatum (..),
+
+  -- * Allegra
+  AllegraTxAuxData (..),
+
+  -- * Alonzo
+  AlonzoTxAuxData (..),
+  mkAlonzoTxAuxData,
+  getAlonzoTxAuxDataScripts,
 )
 where
 
+import Cardano.Ledger.Allegra.TxAuxData (AllegraTxAuxData (..))
+import Cardano.Ledger.Alonzo.TxAuxData (
+  AlonzoTxAuxData (..),
+  getAlonzoTxAuxDataScripts,
+  mkAlonzoTxAuxData,
+ )
+import Cardano.Ledger.Api.Era ()
 import Cardano.Ledger.Core (EraTxAuxData (..))
+import Cardano.Ledger.Shelley.TxAuxData (Metadatum (..), ShelleyTxAuxData (..))