+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/Body.hs
@@ -8,34 +8,69 @@ module Cardano.Ledger.Api.Tx.Body (
   module Cardano.Ledger.Api.Tx.Out,
   -- | Working with Timelock scripts and Plutus scripts
   module Cardano.Ledger.Api.Scripts,
-  EraTxBody (..),
+  EraTxBody (TxBody),
+  mkBasicTxBody,
+  inputsTxBodyL,
+  outputsTxBodyL,
+  feeTxBodyL,
+  withdrawalsTxBodyL,
+  Withdrawals (..),
+  auxDataHashTxBodyL,
+  AuxiliaryDataHash,
+  allInputsTxBodyF,
 
   -- * Shelley Era
-  ShelleyTxBody,
-  ShelleyEraTxBody (..),
+  ShelleyEraTxBody,
+  ttlTxBodyL,
+  updateTxBodyL,
+  updateTxBodyG,
+  certsTxBodyL,
+  certsTxBodyG,
 
   -- * Allegra Era
-  AllegraEraTxBody (..),
+  AllegraEraTxBody,
+  vldtTxBodyL,
+  ValidityInterval (..),
 
   -- * Mary Era
-  MaryEraTxBody (..),
+  MaryEraTxBody,
+  mintTxBodyL,
+  mintValueTxBodyF,
+  mintedTxBodyF,
 
   -- * Alonzo Era
-  AlonzoTxBody,
-  AlonzoEraTxBody (..),
+  AlonzoEraTxBody,
+  collateralInputsTxBodyL,
+  reqSignerHashesTxBodyL,
+  scriptIntegrityHashTxBodyL,
+  networkIdTxBodyL,
 
   -- * Babbage Era
-  BabbageTxBody,
-  BabbageEraTxBody (..),
+  BabbageEraTxBody,
+  sizedOutputsTxBodyL,
+  referenceInputsTxBodyL,
+  totalCollateralTxBodyL,
+  collateralReturnTxBodyL,
+  sizedCollateralReturnTxBodyL,
+  allSizedOutputsTxBodyF,
+
+  -- * Conway Era
+  ConwayEraTxBody,
+  votingProceduresTxBodyL,
+  proposalProceduresTxBodyL,
+  conwayCertsTxBodyL,
 )
 where
 
+import Cardano.Ledger.Address (Withdrawals (..))
 import Cardano.Ledger.Allegra.Core (AllegraEraTxBody (..))
-import Cardano.Ledger.Alonzo.TxBody (AlonzoEraTxBody (..), AlonzoTxBody)
+import Cardano.Ledger.Alonzo.TxBody (AlonzoEraTxBody (..))
+import Cardano.Ledger.Api.Era ()
 import Cardano.Ledger.Api.Scripts
 import Cardano.Ledger.Api.Tx.Out
-import Cardano.Ledger.Babbage.TxBody (BabbageEraTxBody (..), BabbageTxBody)
+import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash)
+import Cardano.Ledger.Babbage.TxBody (BabbageEraTxBody (..))
+import Cardano.Ledger.Conway.Core (ConwayEraTxBody (..))
 import Cardano.Ledger.Core (EraTxBody (..))
 import Cardano.Ledger.Mary.Core (MaryEraTxBody (..))
-import Cardano.Ledger.Shelley (ShelleyTxBody)
 import Cardano.Ledger.Shelley.Core (ShelleyEraTxBody (..))